[shuffle] add shuffle build command
Closes: #9449
